ref(demo-mode): extract to module
