+++ src/sentry/api/endpoints/organization_member/requests/join.py
@@ -11,6 +11,7 @@ from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import region_silo_endpoint
 from sentry.api.bases.organization import OrganizationEndpoint
 from sentry.auth.services.auth import auth_service
+from sentry.demo_mode.utils import is_demo_user
 from sentry.hybridcloud.models.outbox import outbox_context
 from sentry.models.organizationmember import InviteStatus, OrganizationMember
 from sentry.notifications.notifications.organization_request import JoinRequestNotification
@@ -18,7 +19,6 @@ from sentry.notifications.utils.tasks import async_send_notification
 from sentry.signals import join_request_created
 from sentry.types.ratelimit import RateLimit, RateLimitCategory
 from sentry.users.api.parsers.email import AllowedEmailField
-from sentry.utils.demo_mode import is_demo_user
 
 logger = logging.getLogger(__name__)
